Revert "Websocket chats"
