+++ Shynder-app/main.py
@@ -7,98 +7,13 @@ from fastapi.staticfiles import StaticFiles
 from sql.database import SessionLocal, engine
 import bcrypt
 from fastapi import WebSocket, WebSocketDisconnect
-import json
+
 
 models.Base.metadata.create_all(bind=engine)
 app = FastAPI()
-active_users = {"1".encode('utf-8'):db_wrapper.get_user_by_email(SessionLocal(), "bykov.pn@ucu.edu.ua"), "2".encode('utf-8'):db_wrapper.get_user_by_email(SessionLocal(), "sen.pn@ucu.edu.ua"), "3".encode('utf-8'):db_wrapper.get_user_by_email(SessionLocal(), "maria.pn@ucu.edu.ua")}
+active_users = {"1".encode('utf-8'):db_wrapper.get_user_by_email(SessionLocal(), "bykov.pn@ucu.edu.ua")}
 app.mount('/static', StaticFiles(directory='static', html=True), name='static')
-class ConnectionManager:
-    def __init__(self):
-        self.active_sockets = {}
-    async def accept_connection(self, websocket: WebSocket, session_id: str):
-        await websocket.accept()
-        user = active_users[session_id.encode('utf-8')]
-        self.active_sockets[websocket] = (user, session_id)
-        print("Accepted connection")
-        await self.handle_user(websocket)
-    async def close_connection(self, websocket: WebSocket):
-        await websocket.close()
-        del self.active_sockets[websocket]
-    async def send_personal_message(self, message: str, websocket: WebSocket):
-        await websocket.send_text(message)
-    async def handle_user(self, websocket: WebSocket):
-        print("Handling user")
-        try:
-            while True:
-                data = await websocket.receive_text()
-                print(data)
-                message = Message.from_json(data)
-                await self.process_data(message, websocket) 
-        except WebSocketDisconnect:
-            self.close_connection(websocket)
-    
-    async def process_data(self, message: 'Message', websocket: WebSocket):
-        if message.command == "send":
-            receiver = db_wrapper.get_user_by_email(get_db().__next__(), message.receiver)
-            print(receiver.email)
-            print(receiver.id, self.active_sockets[websocket][0].id)
-            match_id = db_wrapper.get_match_id(get_db().__next__(), self.active_sockets[websocket][0].id, receiver.id)
-            if match_id is None:
-                raise ValueError("Match not found")
-            # print(self.active_sockets)
-            print(receiver.email)
-            print(f"Receivers: {[self.active_sockets[socket][0].email for socket in self.active_sockets]}")
-            if receiver.email in [self.active_sockets[socket][0].email for socket in self.active_sockets]:
-                for socket in self.active_sockets:
-                    print(self.active_sockets[socket][0].email, receiver.email)
-                    if self.active_sockets[socket][0].email == receiver.email:
-                        await self.send_personal_message(message.to_json(), socket)
-            try:
-                with open("chat_logs/" + str(match_id) + ".txt", "a") as file:
-                    file.write(message.to_json() + "\n")
-            except FileNotFoundError:
-                file = open("chat_logs/" + str(match_id) + ".txt", "w")
-                file.write(message.to_json() + "\n")
-                file.close()
-        elif message.command == "get_all":
-            match_id = db_wrapper.get_match_id(get_db().__next__(), self.active_sockets[websocket][0].id, db_wrapper.get_user_by_email(get_db().__next__(), message.receiver).id)
-            print(match_id)
-            try:
-                with open("chat_logs/" + str(match_id) + ".txt", "r") as file:
-                    for line in file.readlines():
-                        await self.send_personal_message(line, websocket)
-            except FileNotFoundError:
-                file = open("chat_logs/" + str(match_id) + ".txt", "w")
-                file.close()
-        elif message.command == "get_all_matches":
-            print(self.active_sockets[websocket][0].id)
-            matches = db_wrapper.get_all_matches(get_db().__next__(), self.active_sockets[websocket][0].id)
-            res_matches = []
-            for match in matches:
-                res_matches.append(match.as_dict())
-                res_matches[-1]['command'] = "get_all_matches"
-            await self.send_personal_message(json.dumps(res_matches), websocket)
-        
 
-class Message:
-    def __init__(self, sender:str, receiver:str, messege:str,time, command:str = ""):
-        self.sender = sender
-        self.receiver = receiver
-        self.messege = messege
-        self.command = command
-        self.time = time
-    def to_json(self):
-        encoder = json.encoder.JSONEncoder()
-        return encoder.encode(self.__dict__)
-    @staticmethod
-    def from_json(_json):
-        decoder = json.decoder.JSONDecoder()
-        decoded = decoder.decode(_json)
-        return Message(sender=decoded['sender'], receiver=decoded['receiver'], messege=decoded['messege'], time=decoded['time'], command=decoded['command'])
-
-
-manager = ConnectionManager()
 def get_db():
     db = SessionLocal()
     try:
@@ -231,30 +146,9 @@ async def chats_page():
     with open("./static/chats/chats.html", "r", encoding="utf-8") as file:
         html_con = '\n'.join(file.readlines())
     return html_con
-
-
-@app.get("/match/")
-async def match(user1_id:int, user2_id:int):
-    db = get_db().__next__()
-    db_wrapper.create_match(db, schemas.MatchCreate(user1_id=user1_id, user2_id=user2_id))
-    return {"message": "Success"}
-
-@app.websocket("/chats_websocket/{session_id}")
-async def chats_websocket(websocket: WebSocket, session_id:str):
-    print("Accepted connection")
-    await manager.accept_connection(websocket, session_id)
-
-@app.get("/get_match/")
-async def get_match(match_id:int):
-    db = get_db().__next__()
-    return db_wrapper.get_match(db, match_id)
-
-@app.get("/get_user_by_id/")
-async def get_user_by_id(user_id:int):
-    db = get_db().__next__()
-    return db_wrapper.get_user(db, user_id)
-
+    
+    
 
 if __name__ == "__main__":
     import uvicorn
-    uvicorn.run(app, host="127.0.0.1", port=8000, reload=False)
+    uvicorn.run(app, host="0.0.0.0", port=8000, reload=False)
