+++ Shynder-app/sql/models.py
@@ -17,9 +17,6 @@ class User(Base):
     test_results = Column(String, nullable=False)
     matches = relationship("Match", back_populates="user1", foreign_keys='Match.user1_id')
 
-    def as_dict(self):
-        return {c.name: getattr(self, c.name) for c in self.__table__.columns}
-
 class Match(Base):
     """
     Match model.
@@ -31,6 +28,3 @@ class Match(Base):
     chat_log_file = Column(String, nullable=True)
     user1 = relationship("User", back_populates="matches", foreign_keys=[user1_id])
     user2 = relationship("User", back_populates="matches", foreign_keys=[user2_id])
-
-    def as_dict(self):
-        return {c.name: getattr(self, c.name) for c in self.__table__.columns}
