+++ Shynder-app/static/chats/chats.html
@@ -20,34 +20,34 @@
                 <img src="/static/big_logo.png" class="closeIcon"> 
             </div>
             <div class="menu">
-                <div class="menuItem" onclick="home_click()">Home</div>
-                <div class="menuItem" onclick="profile_click()">Profile</div>
-                <div class="menuItem" onclick="chats_click()">Messages</div>
-                <div class="menuItem" onclick="logout()">Logout</div>
+                <div class="menuItem">Home</div>
+                <div class="menuItem">Profile</div>
+                <div class="menuItem">Messages</div>
+                <div class="menuItem">Logout</div>
             </div> 
             <hr class="line"></hr>
             <div class="chats">
-                <!-- <div class="chat">
-                    <img src="https://www.w3schools.com/howto/img_avatar.png" alt="Avatar" class="chat_avatar">
+                <div class="chat">
+                    <img src="https://www.w3schools.com/howto/img_avatar.png" alt="Avatar" class="avatar">
                     <div class="chat_info">
                         <p class="name">John Doe</p>
                         <p class="last_message">Hey, how are you?</p>
                     </div>
                 </div>
                 <div class="chat">
-                    <img src="https://www.w3schools.com/howto/img_avatar.png" alt="Avatar" class="chat_avatar">
+                    <img src="https://www.w3schools.com/howto/img_avatar.png" alt="Avatar" class="avatar">
                     <div class="chat_info">
-                        <p class="name">John Doe</p>
+                        <h2 class="name">John Doe</h2>
                         <p class="last_message">Hey, how are you?</p>
                     </div>
                 </div>
                 <div class="chat">
-                    <img src="https://www.w3schools.com/howto/img_avatar.png" alt="Avatar" class="chat_avatar">
+                    <img src="https://www.w3schools.com/howto/img_avatar.png" alt="Avatar" class="avatar">
                     <div class="chat_info">
-                        <p class="name">John Doe</p>
+                        <h2 class="name">John Doe</h2>
                         <p class="last_message">Hey, how are you?</p>
                     </div>
-                </div> -->
+                </div>
             </div>
         </div>
         <div class="messages_tab">
@@ -73,5 +73,5 @@
             </div>
         </div>
     </div>
-    <script src="/static/chats/chats.js"></script>
+    <script src="/static/chats/chat.js"></script>
 </body>
